Move spec to correct location
